requirements.txt
fastapi==0.110.0
gradio==4.20.0
huggingface_hub==0.21.3
langchain==0.1.11
langchain-community==0.0.29
langchain-core==0.1.36
langchain-openai==0.1.1
nltk==3.8.1
openai==1.13.3
pip==23.3.1
pydantic==2.6.3
python-dotenv==1.0.1
qianfan==0.3.3
Requests==2.31.0
transformers==4.38.2
websocket_client==1.7.0
zhipuai==2.0.1
chromadb==0.4.14
rapidocr-onnxruntime==1.3.15
pymupdf==1.24.0
unstructured==0.12.6
chromadb==0.4.14
markdown==3.6
spark-ai-python==0.3.15
erniebot==0.5.5




